requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
passlib[bcrypt]
aiofiles
python-multipart
itsdangerous
pillow